sinktools

Module buffered_lazy_sink_source

Available on crate feature std only.

BufferedLazySinkSource, and related items.

Structs

BufferedLazySinkHalf

Sink half of the BufferedLazySinkSource

BufferedLazySinkSource

A buffered lazy sink-source that immediately spawns a task to initialize the connection. Uses tokio channels to forward data between the spawned task and the sink/stream halves.

BufferedLazySourceHalf

Stream half of the BufferedLazySinkSource