sinktools::sink

Struct WithFlatMap

pub struct WithFlatMap<Si, Item, U, St, F> { /* private fields */ }

Sink for the with_flat_map method.

Implementations

impl<Si, Item, U, St, F> WithFlatMap<Si, Item, U, St, F>
where Si: Sink<Item>, F: FnMut(U) -> St, St: Stream<Item = Result<Item, <Si as Sink<Item>>::Error>>,

pub fn get_ref(&self) -> &Si

Acquires a reference to the underlying sink or stream that this combinator is pulling from.

pub fn get_mut(&mut self) -> &mut Si

Acquires a mutable reference to the underlying sink or stream that this combinator is pulling from.

Note that care must be taken to avoid tampering with the state of the sink or stream which may otherwise confuse this combinator.

pub fn get_pin_mut( self: Pin<&mut WithFlatMap<Si, Item, U, St, F>>, ) -> Pin<&mut Si>

Acquires a pinned mutable reference to the underlying sink or stream that this combinator is pulling from.

Note that care must be taken to avoid tampering with the state of the sink or stream which may otherwise confuse this combinator.

pub fn into_inner(self) -> Si

Consumes this combinator, returning the underlying sink or stream.

Note that this may discard intermediate state of this combinator, so care should be taken to avoid losing resources when this is called.