sinktools

Trait ToSinkBuild

pub trait ToSinkBuild {
    // Provided methods
    fn iter_to_sink_build(self) -> SendIterBuild<Self>
       where Self: Sized + Iterator { ... }
    fn stream_to_sink_build(self) -> SendStreamBuild<Self>
       where Self: Sized + Stream { ... }
}

Blanket trait for sending items from Self into a SinkBuild.

Provided Methods

fn iter_to_sink_build(self) -> SendIterBuild<Self>
where Self: Sized + Iterator,

Starts a SinkBuild adaptor chain to send all items from self as an Iterator.

fn stream_to_sink_build(self) -> SendStreamBuild<Self>
where Self: Sized + Stream,

Starts a SinkBuild adaptor chain to send all items from self as a [Stream].